feat(proxy): Extend HTTP reply protocol
The proxy treats a reply from an app depending on
X-Cocaine-HTTP-Proto-Version header presence and value.
Value "1.1" makes proxy treat an empty body chunk (length 0) as
the end of reply.
In case of "1.0" or absence of the header the proxy reads until
closing Cocaine-Stream.

The motivation is to allow applications to send replies to users
as soon as possible and to hold Cocaine Tx stream at the same time.
Apps can do any asynchronous (from the user's point of view) job.
Keeping a Tx stream
is needed to properly counting of current load of a worker.